test/models/WearableDeviceTest.java
/*
package models;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertTrue;

public class WearableDeviceTest {
    private WearableDevice validWearable, invalidWearable;

    @After
    public void tearDown() {
        validWearable = invalidWearable = null;
    }

    @Before
    public void setUp() {
        Manufacturer manufacturer = new Manufacturer("Apple", 500);
        Manufacturer invalidManufacturer = new Manufacturer("ABCDEFGHIJKLMNOPQRSTU", 0);

        // 创建具体的WearableDevice子类对象，这里以SmartWatch为例
        validWearable = new SmartWatch("Apple Watch", 399.99, manufacturer, "123456", "Aluminum", "Medium", "AMOLED");
        invalidWearable = new SmartWatch("Apple Watch Ultra", 19.99, invalidManufacturer, "12345678910", "LongMaterialNameExceedsLimit", "ExtraLargeSize", "Unknown");
    }

    @Test
    public void testGetMaterial() {
        assertEquals("Aluminum", validWearable.getMaterial());
        assertEquals("LongMaterialNameExceeds", invalidWearable.getMaterial());
    }

    @Test
    public void testSetMaterial() {
        validWearable.setMaterial("Titanium");
        assertEquals("Titanium", validWearable.getMaterial());
        validWearable.setMaterial("MaterialNameThatIsTooLong");
        assertEquals("Titanium", validWearable.getMaterial());
    }

    @Test
    public void testGetSize() {
        assertEquals("Medium", validWearable.getSize());
        assertEquals("ExtraLargeSi", invalidWearable.getSize());
    }

    @Test
    public void testSetSize() {
        validWearable.setSize("Large");
        assertEquals("Large", validWearable.getSize());
        validWearable.setSize("SizeTooBigToFit");
        assertEquals("Large", validWearable.getSize());
    }

    @Test
    public void testToString() {
        String expected = "Material: Aluminum, Size: Medium";
        assertTrue(validWearable.toString().contains(expected));
        expected = "Material: LongMaterialNameExceeds, Size: ExtraLargeSi";
        assertTrue(invalidWearable.toString().contains(expected));
    }
}
*/
